SPGAmerica/requirements.txt
asgiref
aspose-words
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-cleanup==6.0.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
